working on animation

+++ 37MonsterHunter/engine/RandomColor.cpp
@@ -152,15 +152,15 @@ int RandomColor::pickBrightness( int s, const ColorInfo& info, Luminosity lumino
 
     switch (luminosity) {
         case Dark:
-            bRange[1] = bRange[0] + std::min(bRangeSize * 0.3f, 30.0f);
+            bRange[1] = bRange[0] + static_cast<int>(std::min(bRangeSize * 0.3f, 30.0f));
             break;
 
         case Light:
-            bRange[0] = bRange[1] - std::min(bRangeSize * 0.3f, 15.0f);
+            bRange[0] = bRange[1] - static_cast<int>(std::min(bRangeSize * 0.3f, 15.0f));
             break;
 
         case Bright:
-            bRange[0] = bRange[1] - std::min(bRangeSize * 0.3f, 10.0f);
+            bRange[0] = bRange[1] - static_cast<int>(std::min(bRangeSize * 0.3f, 10.0f));
             break;
 
         case Normal:
